+++ opening-hours-syntax/src/parser.rs
@@ -243,32 +243,32 @@ fn build_time_selector(pair: Pair<Rule>) -> Result<Vec<ts::TimeSpan>> {
 fn build_timespan(pair: Pair<Rule>) -> Result<ts::TimeSpan> {
     assert_eq!(pair.as_rule(), Rule::timespan);
     let mut pairs = pair.into_inner();
-
     let start = build_time(pairs.next().expect("empty timespan"))?;
 
-    let end = match pairs.next() {
+    let (open_end, end) = match pairs.next() {
         None => {
-            // TODO: opening_hours.js handles this better: it will set the
-            //       state to unknown and add a warning comment.
-            ts::Time::Fixed(ExtendedTime::new(24, 0).unwrap())
-        }
-        Some(pair) if pair.as_rule() == Rule::timespan_plus => {
             return Err(Error::Unsupported("point in time"));
         }
-        Some(pair) => build_extended_time(pair)?,
+        Some(pair) if pair.as_rule() == Rule::timespan_plus => {
+            // TODO: opening_hours.js handles this better: it will set the
+            //       state to unknown and add a warning comment.
+            (true, ts::Time::Fixed(ExtendedTime::new(24, 0).unwrap()))
+        }
+        Some(pair) => (false, build_extended_time(pair)?),
     };
 
     let (open_end, repeats) = match pairs.peek().map(|x| x.as_rule()) {
-        None => (false, None),
+        None => (open_end, None),
         Some(Rule::timespan_plus) => (true, None),
-        Some(Rule::minute) => (false, Some(build_minute(pairs.next().unwrap()))),
+        Some(Rule::minute) => (open_end, Some(build_minute(pairs.next().unwrap()))),
         Some(Rule::hour_minutes) => (
-            false,
+            open_end,
             Some(build_hour_minutes_as_duration(pairs.next().unwrap())),
         ),
         Some(other) => unexpected_token(other, Rule::timespan),
     };
 
+    assert!(pairs.next().is_none());
     Ok(ts::TimeSpan { range: start..end, repeats, open_end })
 }
 
@@ -691,12 +691,11 @@ fn build_hour_minutes(pair: Pair<Rule>) -> Result<ExtendedTime> {
     assert_eq!(pair.as_rule(), Rule::hour_minutes);
     let mut pairs = pair.into_inner();
 
-    let hour = pairs
-        .next()
-        .expect("missing hour")
-        .as_str()
-        .parse()
-        .expect("invalid hour");
+    let Some(hour_rule) = pairs.next() else {
+        return Ok(ExtendedTime::new(24, 0).unwrap());
+    };
+
+    let hour = hour_rule.as_str().parse().expect("invalid hour");
 
     let minutes = pairs
         .next()